feat: 클래스 추출
