+++ README.md
@@ -19,7 +19,8 @@
 - [x] 메소드 추출
   - createQueryForInsert(), createQueryForUpdate()
   - setValuesForInsert(User, PreparedStatement), setValuesForUpdate(User, PreparedStatement)
-- [ ] 클래스 추출
+- [x] 클래스 추출
+  - `InsertJdbcTemplate`, `UpdateJdbcTemplate`
 - [ ] 템플릿 메소드 패턴 적용
 - [ ] 도메인 의존도 제거
 - [ ] 클래스 추출
